// src/adapters/ravepage/api-client/models/ShowcaseBackgroundOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShowcaseBackgroundOut = {
    file_size?: number;
    kind?: 'image' | 'video';
    media_upload_id?: string;
    mime_type?: string;
    url?: string;
};