// src/adapters/ravepage/api-client/models/ShowcaseCollaboratorUpdateIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShowcaseCollaboratorUpdateIn = {
    /**
     * MinGroupRole is the new minimum group role.
     */
    min_group_role?: string;
    /**
     * Role is the new grant role.
     */
    role?: 'owner' | 'editor' | 'viewer';
};