// src/adapters/ravepage/api-client/models/ShowcasePageListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShowcasePageListItem = {
    created_at?: string;
    display_name?: string;
    id?: string;
    is_published?: boolean;
    owner_id?: string;
    owner_type?: string;
    owner_user_username?: string;
    short_code?: string;
    slug?: string;
    type?: string;
    updated_at?: string;
    user_role?: string;
    visibility?: 'public' | 'unlisted' | 'private';
};